fix(http-client-csharp): make emitter.ts and lib/utils.ts browser-bundleable
Moves the Node-only execCSharpGenerator/execAsync helpers (which use
child_process) out of lib/utils.ts and into a new lib/exec-utils.ts that
is only imported from emit-generate.ts (the Node-only generation entry
point). The browser variant (emit-generate.browser.ts) does not import
exec-utils, so child_process is no longer pulled into the browser bundle.

Also replaces `import { resolve } from "path"` in emitter.ts with the
cross-platform `resolvePath` utility from @typespec/compiler so that
emitter.ts no longer has a direct dependency on the Node `path` module.

This fixes the "Upload playground bundle" pipeline step that was failing
with:
  emitter.js: Could not resolve "path"
  lib/utils.js: Could not resolve "child_process"
