+++ packages/http-client-csharp/emitter/test/Unit/validate-dotnet-sdk.test.ts
@@ -4,7 +4,7 @@ import { Diagnostic, Program } from "@typespec/compiler";
 import { TestHost } from "@typespec/compiler/testing";
 import { strictEqual } from "assert";
 import { beforeEach, describe, expect, it, Mock, vi } from "vitest";
-import { execAsync } from "../../src/lib/utils.js";
+import { execAsync } from "../../src/lib/exec-utils.js";
 import {
   createCSharpSdkContext,
   createEmitterContext,
@@ -33,7 +33,7 @@ describe("Test _validateDotNetSdk", () => {
     );
     // Restore all mocks before each test
     vi.restoreAllMocks();
-    vi.mock("../../src/lib/utils.js", () => ({
+    vi.mock("../../src/lib/exec-utils.js", () => ({
       execCSharpGenerator: vi.fn(),
       execAsync: vi.fn(),
     }));
